+++ changelog.md
@@ -1,2 +1,3 @@
 - Youtube playables (games grid) toggle
 - Hide sub count on channel pages/videos toggle
+- Hide top bar avatar toggle